+++ spec/src/general/music/isRoughSpec.ts
@@ -1,9 +1,10 @@
 import { computeIsRough } from "../../../../src/general"
+import { Integer } from "../../../../src/general/math"
 
 describe("isRough", () => {
     it("works for integers", () => {
-        const integer = 221
-        const roughness = 11
+        const integer = 221 as Integer
+        const roughness = 11 as Integer
 
         const actual = computeIsRough(integer, roughness)
 
@@ -11,8 +12,8 @@ describe("isRough", () => {
     })
 
     it("works for integers, example where it is false", () => {
-        const integer = 33
-        const roughness = 5
+        const integer = 33 as Integer
+        const roughness = 5 as Integer
 
         const actual = computeIsRough(integer, roughness)
 
@@ -20,8 +21,8 @@ describe("isRough", () => {
     })
 
     it("1 is always rough", () => {
-        const integer = 1
-        const roughness = 5
+        const integer = 1 as Integer
+        const roughness = 5 as Integer
 
         const actual = computeIsRough(integer, roughness)
 
@@ -29,8 +30,8 @@ describe("isRough", () => {
     })
 
     it("trying to find whatever edge cases causes this insanely faster implementation to disagree with the old way", () => {
-        const integer = 10
-        const roughness = 11
+        const integer = 10 as Integer
+        const roughness = 11 as Integer
 
         const actual = computeIsRough(integer, roughness)
 
@@ -38,8 +39,8 @@ describe("isRough", () => {
     })
 
     it("even more examples", () => {
-        const integer = 11
-        const roughness = 11
+        const integer = 11 as Integer
+        const roughness = 11 as Integer
 
         const actual = computeIsRough(integer, roughness)
